\name{envdata}
\alias{envdata}
\docType{data}
\title{
Survey data of Environmental factors for pine-oak mixed forests
}
\description{
Environmental factors of pine-oak codominant mixed forests in the Qinling Mountains
}
\usage{data("envdata")}
\format{
  A data frame with 20 observations on the following 16 variables.
  \describe{
    \item{\code{plot.id}}{Identification number of sample plots}
    \item{\code{Ele}}{Elevation}
    \item{\code{SLP}}{Slope position}
    \item{\code{Asp}}{Slope aspect}
    \item{\code{SLG}}{Slope gradient}
    \item{\code{TN}}{Total nitrogen}
    \item{\code{TP}}{Total phosphorus}
    \item{\code{TK}}{Total potassium }
    \item{\code{OM}}{Organic matter}
    \item{\code{pH}}{pH value}
    \item{\code{AN}}{Available nitrogen}
    \item{\code{AP}}{Available phosphorus }
    \item{\code{AK}}{Available potassium }
    \item{\code{BA}}{Base area}
    \item{\code{AD}}{Average tree DBH }
    \item{\code{AH}}{Average tree height}
  }
}
\details{
Twenty 20*20 m plots were established for the pine-oak codominant
mixed forests within the Qinling National Forest Ecosystem Research
Station (Huoditang forest)
}
\references{
Chai ZZ, Wang DX. 2016. Environmental influences on the successful regeneration of pine-oak mixed forests in the Qinling Mountains, China. Scandinavian Journal of Forest Research,31(4):368-381.DOI:10.1080/02827581.2015.1062912
}
\examples{
data(envdata)
envdata
}
