\name{herbdata}
\alias{herbdata}
\docType{data}
\title{
Survey data of undergrowth herb cluster for pine-oak mixed forests
}
\description{
Undergrowth herb data of pine-oak codominant mixed forests in the Qinling Mountains
}
\usage{data("herbdata")}
\format{
  A data frame with 229 observations on the following 4 variables.
  \describe{
    \item{\code{plot.id}}{Identification number of sample plots}
    \item{\code{spe.id}}{Identification number of tree species}
    \item{\code{coverage}}{Coverage of herb species}
    \item{\code{abundance}}{Abundance of herb species}
  }
}
\details{
Twenty 20*20 m plots were established for the pine-oak codominant
mixed forests within the Qinling
National Forest Ecosystem Research Station (Huoditang
forest)
}
\references{
Chai ZZ, Wang DX. 2016. Environmental influences on the successful regeneration of pine-oak mixed forests in the Qinling Mountains, China. Scandinavian Journal of Forest Research,31(4):368-381.DOI:10.1080/02827581.2015.1062912
}
\examples{
data(herbdata)
herbdata
}

