\name{indexSystem}
\alias{indexSystem}
\docType{data}
\title{
The national forest health evaluation system
}
\description{
The national forest health evaluation system
}
\usage{data("indexSystem")}
\format{
  A data frame with 21 observations on the following 9 variables.
  \describe{
    \item{\code{Criterion}}{a factor with levels \code{} \code{b1} \code{b2} \code{b3} \code{b4} \code{b5} \code{b6} \code{b7}}
    \item{\code{WeightB}}{a numeric vector}
    \item{\code{Index}}{a factor with levels \code{c1} \code{c10} \code{c11} \code{c12} \code{c13} \code{c14} \code{c15*} \code{c16*} \code{c17*} \code{c18*} \code{c19*} \code{c2} \code{c20*} \code{c21*} \code{c3} \code{c4} \code{c5} \code{c6} \code{c7} \code{c8*} \code{c9}}
    \item{\code{WeightC}}{a numeric vector}
    \item{\code{Mode}}{a factor with levels \code{down} \code{mid} \code{up}}
    \item{\code{L}}{a numeric vector}
    \item{\code{U}}{a numeric vector}
    \item{\code{O1}}{a numeric vector}
    \item{\code{O2}}{a numeric vector}
  }
}
\details{
ID see legend
}
\source{
Zhang HR, and Lei XD. 2014. Health management techniques for typical forest types. Beijing: Publishing house of forestry, China.
Chai ZZ.2016.National forest health evaluation system at the forest stand level in China.http://www.forest-soil.net/Upload/ueditor/file/20160514/1463221795322480.pdf
}
\references{
Zhang HR, and Lei XD. 2014. Health management techniques for typical forest types. Beijing: Publishing house of forestry, China
Chai ZZ.2016.National forest health evaluation system at the forest stand level in China\url{http://www.forest-soil.net/Upload/ueditor/file/20160514/1463221795322480.pdf}
}
\examples{
data(indexSystem)
}

