\name{testIndex}
\alias{testIndex}
\docType{data}
\title{
Test Indices for forest health evaluation system
}
\description{
Example data of test Indices for forest health evaluation system
}
\usage{data("testIndex")}
\format{
  A data frame with 30 observations on the following 21 variables.
  \describe{
    \item{C1}{Tree diversity index}
  \item{C2}{Shrub diversity index}
  \item{C3}{Herb diversity index}
  \item{C4}{Storey quantity}
  \item{C5}{Tree size diversity index}
  \item{C6}{Age structure}
  \item{C7}{Canopy closure}
  \item{C8}{Average volume increment per hectare}
  \item{C9}{Bio-disaster ratio}
  \item{C10}{Non bio-disaster ratio}
  \item{C11}{Crown length ratio}
  \item{C12}{Crown fade ratio}
  \item{C13}{Foliage falling ratio}
  \item{C14}{Regeneration density}
  \item{C15}{Soil depth}
  \item{C16}{Soil density}
  \item{C17}{Soil pH}
  \item{C18}{Soil organic matter}
  \item{C19}{Soil total nitrogen}
  \item{C20}{Soil available phosphorus}
  \item{C21}{Soil available potassium}
  }
}
\examples{
data(testIndex)
testIndex
}

