\name{treedata}
\alias{treedata}
\docType{data}
\title{
Survey data of individual trees for pine-oak mixed forests
}
\description{
Tree data of pine-oak codominant mixed forests in the Qinling Mountains
}
\usage{data("treedata")}
\format{
  A data frame with 1054 observations on the following 5 variables.
  \describe{
    \item{\code{plot.id}}{Identification number of sample plots}
    \item{\code{tree.id}}{Identification number of trees}
    \item{\code{spe.id}}{Identification number of tree species}
    \item{\code{dbh.cm}}{Diameter at breast height of trees}
    \item{\code{ht.m}}{Tree height of trees}
  }
}
\details{
Twenty 20*20 m plots were established for the pine-oak codominant
mixed forests within the Qinling
National Forest Ecosystem Research Station (Huoditang forest)
}
\references{
Chai ZZ, Wang DX. 2016. Environmental influences on the successful regeneration of pine-oak mixed forests in the Qinling Mountains, China. Scandinavian Journal of Forest Research,31(4):368-381.DOI:10.1080/02827581.2015.1062912
}
\examples{
data(treedata)
treedata
}

