\name{testIndex}
\alias{testIndex}
\docType{data}
\title{
Test Indices for forest health evaluation system
}
\description{
Example data of test Indices for forest health evaluation system
}
\usage{data("testIndex")}
\format{
  A data frame with 30 observations on the following 21 variables.
  \describe{
    \item{c1}{Tree diversity index}
  \item{c2}{Shrub diversity index}
  \item{c3}{Herb diversity index}
  \item{c4}{Storey quantity}
  \item{c5}{Tree size diversity index}
  \item{c6}{Age structure}
  \item{c7}{Canopy closure}
  \item{c8}{Average volume increment per hectare}
  \item{c9}{Bio-disaster ratio}
  \item{c10}{Non bio-disaster ratio}
  \item{c11}{Crown length ratio}
  \item{c12}{Crown fade ratio}
  \item{c13}{Foliage falling ratio}
  \item{c14}{Regeneration density}
  \item{c15}{Soil depth}
  \item{c16}{Soil density}
  \item{c17}{Soil pH}
  \item{c18}{Soil organic matter}
  \item{c19}{Soil total nitrogen}
  \item{c20}{Soil available phosphorus}
  \item{c21}{Soil available potassium}
  }
}
\examples{
data(testIndex)
testIndex
}

